requirements.txt
fastapi>=0.115.0,<1.0.0
google-auth>=2.0.0,<3.0.0
google-genai>=1.9.0
gspread>=6.0.0,<7.0.0
uvicorn>=0.34.0,<1.0.0